% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_codes.R
\name{fetch_codes}
\alias{fetch_codes}
\title{Retrieve state and district codes}
\usage{
fetch_codes(regions = c("states", "state", "districts", "district"))
}
\arguments{
\item{regions}{The region breakdown for the map, can be one of
(\code{"states"}, \code{"state"}, \code{"districts"}, \code{"district"}).
The default is \code{"states"}.}
}
\value{
An data frame of codes of the desired \code{regions}.
}
\description{
Retrieve state and district codes
}
\examples{
str(fetch_codes())

state_codes <- fetch_codes()

district_codes <- fetch_codes(regions = "districts")
}
